+++ psgscoring/__init__.py
@@ -82,7 +82,7 @@ from .utils import (
     safe_r,
 )
 
-__version__ = "0.2.962"
+__version__ = "0.2.963"
 __all__ = [
     # Master
     "run_pneumo_analysis",